refactor(api): move business logic out of CoW Swap route into CowSwapService
- Add calculateOrderStatus() to CowSwapService (formerly calculateTrafficLightStatus in route)
- Add processOrders() to CowSwapService for raw order transformation
- Slim down route handler to delegate processing to the service
- Remove local interfaces and function definitions from route.ts
